PeerProtocolSpec, PrefixFlowSpec: move to peer.protocol package, fix some test names
